+++ src/modules/qahw/qahw-sink-extn.c
@@ -522,7 +522,7 @@ static void pa_qahw_sink_set_matrix_param(DBusConnection *conn, DBusMessage *msg
     dbus_message_iter_recurse(&struct_i, &array_i);
     dbus_message_iter_get_fixed_array(&array_i, addr_value, &n_elements);
 
-    if (n_elements > AUDIO_CHANNEL_COUNT_MAX) {
+    if (n_elements > (int)AUDIO_CHANNEL_COUNT_MAX) {
         pa_dbus_send_error(conn, msg, DBUS_ERROR_FAILED , "set_matrix_param failed, unsupported channel count %d", n_elements);
         dbus_error_free(&error);
     }
@@ -537,7 +537,7 @@ static void pa_qahw_sink_set_matrix_param(DBusConnection *conn, DBusMessage *msg
     dbus_message_iter_recurse(&struct_i, &array_i);
     dbus_message_iter_get_fixed_array(&array_i, addr_value, &n_elements);
 
-    if (n_elements > AUDIO_CHANNEL_COUNT_MAX) {
+    if (n_elements > (int)AUDIO_CHANNEL_COUNT_MAX) {
         pa_dbus_send_error(conn, msg, DBUS_ERROR_FAILED , "set_matrix_param failed, unsupported channel count %d", n_elements);
         dbus_error_free(&error);
         return;
@@ -562,7 +562,7 @@ static void pa_qahw_sink_set_matrix_param(DBusConnection *conn, DBusMessage *msg
 
         dbus_message_iter_recurse(&array_i, &array_ii);
         dbus_message_iter_get_fixed_array(&array_ii, addr_value, &n_elements);
-        if (n_elements > AUDIO_CHANNEL_COUNT_MAX) {
+        if (n_elements > (int)AUDIO_CHANNEL_COUNT_MAX) {
             pa_dbus_send_error(conn, msg, DBUS_ERROR_FAILED , "set_matrix_param failed, unsupported channel count %d for mixer coeffcients", n_elements);
             dbus_error_free(&error);
             return;
